Moved the general search methods to a separate file
